+++ api/src/main/java/org/allaymc/api/entity/effect/type/EffectAbsorptionType.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.entity.effect.type;
 
 import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.attribute.AttributeType;
 import org.allaymc.api.entity.component.attribute.EntityAttributeComponent;
 import org.allaymc.api.entity.effect.AbstractEffectType;
 import org.allaymc.api.entity.effect.EffectInstance;
@@ -16,14 +17,14 @@ public class EffectAbsorptionType extends AbstractEffectType {
 
     @Override
     public void onAdd(Entity entity, EffectInstance effectInstance) {
-        if (entity instanceof EntityAttributeComponent attributeComponent && attributeComponent.supportAbsorption()) {
+        if (entity instanceof EntityAttributeComponent attributeComponent && attributeComponent.supportAttribute(AttributeType.ABSORPTION)) {
             attributeComponent.setAbsorption(effectInstance.getLevel() * 4);
         }
     }
 
     @Override
     public void onRemove(Entity entity, EffectInstance effectInstance) {
-        if (entity instanceof EntityAttributeComponent attributeComponent && attributeComponent.supportAbsorption()) {
+        if (entity instanceof EntityAttributeComponent attributeComponent && attributeComponent.supportAttribute(AttributeType.ABSORPTION)) {
             attributeComponent.setAbsorption(0);
         }
     }