+++ api/src/main/java/org/allaymc/api/entity/effect/type/EffectHealthBoostType.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.entity.effect.type;
 
 import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.attribute.AttributeType;
 import org.allaymc.api.entity.component.attribute.EntityAttributeComponent;
 import org.allaymc.api.entity.effect.AbstractEffectType;
 import org.allaymc.api.entity.effect.EffectInstance;
@@ -16,14 +17,16 @@ public class EffectHealthBoostType extends AbstractEffectType {
 
     @Override
     public void onAdd(Entity entity, EffectInstance effectInstance) {
-        if (!(entity instanceof EntityAttributeComponent component) || !component.supportHealth()) return;
+        if (!(entity instanceof EntityAttributeComponent component) || !component.supportAttribute(AttributeType.HEALTH))
+            return;
         var level = effectInstance.getLevel();
         component.setMaxHealth(component.getMaxHealth() + (level * 4));
     }
 
     @Override
     public void onRemove(Entity entity, EffectInstance effectInstance) {
-        if (!(entity instanceof EntityAttributeComponent component) || !component.supportHealth()) return;
+        if (!(entity instanceof EntityAttributeComponent component) || !component.supportAttribute(AttributeType.HEALTH))
+            return;
         var level = effectInstance.getLevel();
         component.setMaxHealth(component.getMaxHealth() - (level * 4));
         if (component.getHealth() > component.getMaxHealth()) {