+++ server/src/main/java/org/allaymc/server/block/component/BlockTntBaseComponentImpl.java
@@ -1,0 +1,21 @@
+package org.allaymc.server.block.component;
+
+import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.dto.PlayerInteractInfo;
+import org.allaymc.api.block.type.BlockType;
+import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.world.Dimension;
+
+/**
+ * @author daoge_cmd
+ */
+public class BlockTntBaseComponentImpl extends BlockBaseComponentImpl {
+    public BlockTntBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+        super(blockType);
+    }
+
+    @Override
+    public boolean onInteract(ItemStack itemStack, Dimension dimension, PlayerInteractInfo interactInfo) {
+        super.onInteract(itemStack, dimension, interactInfo);
+    }
+}